Change name and location of util function
